Add input managers, update camera, and graphics settings
- Introduce Keyboard and MouseInputManager prefabs and references in GameManager.
- Update CameraManager to use int range and adjust camera offset/rotation.
- Rename ZoomInputManager delegate to ZoomInputHandler.
- Update main camera tag, GameManager prefab references, and graphics settings for URP and lighting.
- Enhance post-processing profile, update planet/sun materials, and add debug input axes.

+++ Assets/Scripts/Managers/CameraManager.cs
@@ -108,7 +108,7 @@ namespace SpaceRTS.Managers
 		/// </summary>
 		/// <param name="cameraMode">The desired camera mode.</param>
 		/// <param name="range">The range for the camera movement bounds.</param>
-		public void SetCamera(CameraMode cameraMode, float range)
+		public void SetCamera(CameraMode cameraMode, int range)
         {
 			this.ResolveCamera();
